githubemoji can now update itself with latest emoji
Github changing things up again, but this time I'm ready :). GithubEmoji
extension can be run directly and will generate a new emoji regex and
emoji name to asset link mappying.  This is done with the Github API. No
more manual updating.  If the file is behind, and a user NEEDS the
latest, the extension can be run in offline mode and will grab the
latest from github.  No oauth support, so you will be limited to 60
calls.  Updated emoji-gen to just spit out a markdown file for visual
inspection that emoji are converted.