Don't worry about default packages not being present
- Originally I was planning on future when Python Markdown 3 was not
going to ship the default packages by default.  This would create a
situation where the default packages might not be present.
Realistically we already can the dependency of for Python Markdown < 3,
and it sounds like 3 is going to have more differences than I can plan
for anyways, so let’s move forward with the 2 assumptions: that default
packages are present etc.  When 3 releases, we I probably release a
non-backwards compatible, 3 compliant version then.