Scss changes (#97)
Mainly lint changes. Also ensure line highlight is inline style. Recent changes for clipboard appear to make it hard for block spans to be used for line highlight as the block element can't get the full width of the parent to size the highlight line accordingly. It is possible to fix this with display: sticky by creating a block region that is the size of the parent's width and make it sticky when scrolling, but sticky support across browsers is bad...really bad. It worked in Chrome, but Firefox made an infinitely scrollable code block, and others don't support it at all. Doing this with JavaScript is expensive and not worth it. Should all of the code blocks be reverted in the current theme to work the old way, and then overlay the clipboard icon over the code from an inserted position before the code block instead of in? Maybe. I will probably raise an issue on Material at some point. Not really an issue for me as I don't use highlight line in the real world applications except to demonstrate it is possible in this documentation.