More flexible spellcheck ignore
Aspell is limited to ignoring a tag name and that's it.  This requires you wrap content in nonsense tags like `nospell`.  This pushes the HTML parsing into the wrapper script allowing us to handle what tags get ignored with things like `tagname.class` etc.  This allows us much more flexibility in targetting known  document content that is problematic for the spellchecker and ignoring it like commit hashes for magiclink.