Use script insert for arithmatex
Script format can utilze the `MathJax_Preview` class to insert previews with script mode.  This makes it so script mode (the most reliable insertion format) no longer has the con of no preview making it the best choice.  Allowing MathJax to scan for text for conversion is no longer necessary and obsoletes a number of our options. Ref #171.