fix path conversion protocol (#408)
* fix path conversion protocol

On Windows, we will use the notation `///c:/path` when converting an an absolute drive path to a URL.  `//` means use the current default protocol.  This isn't necessarily correct in the way we are using it.  Instead we should use either apply the `file://` scheme  to all our converted links or  strip the default assumed protocol. For now we are going to strip the assumed default protocol.

* Add test for windows relative paths

Add a test that covers windows paths for relative links.  Also consolidate windows specific path 2 url conversion to one place.