Update variable to be `def_non_std_image_path` (#1228)
When using `twemoji` I noticed if you pass through a name that's not unicode in the index you get unboundlocalerror for `def_non_std_image_path`. I believe everything in the current index is unicode so this likely never comes up but it seems to be a typo since `def_image_path` is defined twice consecutively.