Rename mermaid fences to diagram
The Material theme we use will, one day in the future, automatically
handle mermaid diagrams if they see a code block with `mermaid`.

While this behavior is appreciated, we currently have our own
implementation and styling. We are not currently ready to switch over
and want to avoid such a change until we are ready (if we are ever
ready).