Allow only guessing language for inline/block code
There was no way to specify the type of code block to be guessed
language, for example, to only guess language for code blocks but not
inline code, the length of which is usually insufficient for
satisfactory accuracy.

Allow 'inline' and 'block' to be specified for guess_lang in addition to
boolean values.

Fixes #1696.