Add back in fenced code reversal
The idea is that if we code is contained within a an atomic block or
span that fenced code shouldn't have triggered anyways. This puts it
back to the original source to be handled line inline content.

Reverting in the case of span is probably debatable, but from a pure
HTML perspective, it is the right call. We could potential relax this
in the required later, but atomic should always do this.