Docs: Include MathJax menu CSS by default in custom CSS
MathJax menu CSS gets wiped out by default in instant mode. This
destroys menu functionality and rendering. Additionally, MathJax
provides no on-demand way to regenerate the CSS and insert it into the
head.

To fix this with instant mode, we have no choice but to generate the CSS
ahead of time and include it in our static CSS so that it is always
available.

Fixes #2618