Wrap Pygments lines with <span> to allow for better styling
Inline was always difficult to get line highlights to span the entire
scrollable code block. By ensuring lines are wrapped with a span, code
can be set to `display: grid` allowing for better line highlights.

This should be a benign change and behaves like the `line_span` option,
just without adding the IDs.